edit : headless ui 구현을 위한 수정 작업, PurposeCard를 진행
